Added GCD in C++
The logic of this program is simple.
In this program, the smaller integer between n1 and n2 is stored in n2. Then the loop is iterated from i = 1 to i <= n2 and in each iteration, the value of i is increased by 1.
If both numbers are divisible by i then, that number is stored in variable gcd.
This process is repeated in each iteration. When the iteration is finished, GCD will be stored in variable gcd.