Create DNF_Sort
This Algorithm is also known as Dutch National Flag Sort. The Algorithm can sort an array whose elements have only three different values. This code is for (0,1,2) values. However the Algorithm can be easily modified for any three values. Example of Array that it can sort is : Size of Array(n): 10 , Array : 0,2,2,1,0,1,2,1,0,1